refactor: Refactor linear-probing
Looked at the answer for the assignment and found a few edge cases that were handled way better than I did. Decided to go with the answer from the video so I refactored it and typed it out by hand to make sure I understood everything as it was coded.